Enable FineIBT support
This creates stubs on function prologues to enforce a finer-grained
CFI policy on top of CET/IBT. Because control-flow needs to go through
ENDBR instructions when IBT is enabled, the stubs cannot be bypassed.
This way, before the indirect call takes place, a hash respective to
the function pointer prototype is stored in R11. Then, in the stub, it
is verefied for a match with the function's prototype. If these don't
match, then execution reaches and UD2 instruction. If the hashes match,
the function is executed.